Hide the story text until the user has filled all of the text fields
Adds a "generate" button that is active once all of the text input
fields have been filled in. This button hides the input controls and
displays the completed story text. A reset button can be clicked to hide
the story and re-display the input controls, which are set to their
default values.